+++ diskover-web/src/diskover/Diskover.php
@@ -26,7 +26,7 @@ error_reporting(E_ALL ^ E_NOTICE);
 /* Start Globals */
 
 // diskover-web version
-$VERSION = '2.0-rc.3 community edition (ce)';
+$VERSION = '2.0-rc.4 community edition (ce)';
 
 // array of constants names required in Constants.php config file
 $config_const = array(
@@ -42,7 +42,8 @@ $config_const = array(
     'SEARCH_RESULTS',
     'SIZE_FIELD',
     'FILE_TYPES',
-    'EXTRA_FIELDS'
+    'EXTRA_FIELDS',
+    'MAX_INDEX'
 );
 
 // sanity check of Constants.php config
@@ -58,7 +59,7 @@ if (!isset($_SESSION['sanitycheck'])) {
     $_SESSION['sanitycheck'] = true;
 }
 
-$esIndex = $path = $toppath = $indices_all = $all_index_info = $indices_sorted = $completed_indices = 
+$esIndex = $path = $toppath = $es_index_info = $all_index_info = $indices_sorted = $completed_indices = 
     $latest_completed_index = $fields = $indexinfo_updatetime = $index_starttimes = $es_responsetime = 
     $filter = $time = $use_count = $show_files = $maxdepth = $sizefield = null;
 
@@ -77,6 +78,9 @@ setd3Vars();
 
 // time in seconds for index info session var to expire and force reload indices
 $indexinfo_expiretime = 600;
+
+// time in seconds to check Elasticsearch for new index info
+$esindex_checktime = 10;
 
 // timezone
 // check for env var TZ
@@ -112,25 +116,103 @@ class ESClient
     function getIndicesInfoCurl()
     {
         // Get index info using curl
+
+        // fields to not display
+        $field_exclusions = [
+            'path',
+            'total',
+            'used',
+            'free',
+            'available',
+            'file_size',
+            'file_size_du',
+            'file_count',
+            'dir_count',
+            'start_at',
+            'end_at',
+            'crawl_time',
+            'diskover_ver'
+        ];
+
         $indices_curl_info = curl_es('/diskover-*?pretty');
-        return $indices_curl_info;
+        $indices_curl_info_data = array();
+        // get the data we care about
+        foreach ($indices_curl_info as $key => $val) {
+            $fields = array();
+            foreach ($val['mappings']['properties'] as $fieldname => $fieldtype) {
+                if (!in_array($fieldname, $field_exclusions)) {
+                    $fields[] = $fieldname;
+                }
+            }
+            $indices_curl_info_data[$key] = [
+                'uuid' => $val['settings']['index']['uuid'],
+                'creation_date' => $val['settings']['index']['creation_date'],
+                'fields' => $fields
+            ];
+        }
+        return $indices_curl_info_data;
     }
 
     function getIndicesInfoCat()
     {
         // Get index info using cat
-        $indices_cat_info = $this->client->cat()->indices(array('index' => 'diskover-*'));
-        $indices_cat_info_indexkeys = array();
-        foreach ($indices_cat_info as $i) {
-            $indices_cat_info_indexkeys[$i['index']] = $i;
+        // cat array is sorted by creation date, limit the max number of indices to load
+        $maxindex = getCookie('maxindex');
+        if ($maxindex == '') {
+            $maxindex = Constants::MAX_INDEX;
+            createCookie('maxindex', $maxindex);
         }
-        return $indices_cat_info_indexkeys;
+        // only get diskover indices and sort by creation date
+        $params = array(
+            'index' => 'diskover-*',
+            's' => 'creation.date'
+        );
+        $indices_cat_info = $this->client->cat()->indices($params);
+        $_SESSION['total_indices'] = sizeof($indices_cat_info);
+        // slice array to only limited number of index
+        $indices_cat_info_limited = array_slice($indices_cat_info, -$maxindex, $maxindex, true);
+        $indices_cat_info_data = array();
+        foreach ($indices_cat_info_limited as $i) {
+            $indices_cat_info_data[$i['index']] = [
+                'docs_count' => $i['docs.count'],
+                'store_size' => $i['store.size']
+            ];
+        }
+        return $indices_cat_info_data;
     }
 
     function refreshIndices()
     {
         // Refresh diskover indices
         $this->client->indices()->refresh(array('index' => 'diskover-*'));
+    }
+
+    function getIndexInfo()
+    {
+        global $esindex_checktime;
+        // get index info from ES
+        // only get new index info every 10 seconds
+        if (isset($_GET['reloadindices']) || !isset($_SESSION['es_index_info_time']) || time() - $_SESSION['es_index_info_time'] > $esindex_checktime) {
+            $this->refreshIndices();
+            $indices_info_curl = $this->getIndicesInfoCurl();
+            $indices_info_cat = $this->getIndicesInfoCat();
+            $es_index_info = array();
+            foreach ($indices_info_cat as $index => $val) {
+                $val2 = $indices_info_curl[$index];
+                $es_index_info[$index] = $val + $val2;
+            }
+            $_SESSION['es_index_info'] = $es_index_info;
+            $_SESSION['es_index_info_time'] = time();
+        } else {
+            $es_index_info = $_SESSION['es_index_info'];
+        }
+        return $es_index_info;
+    }
+
+    function getTotalIndices()
+    {
+        // Return total number of indices
+        return $_SESSION['total_indices'];
     }
 }
 
@@ -248,7 +330,7 @@ function setPaths()
 // Set vars and cookies for index, path, etc
 function init()
 {
-    global $timezone, $esIndex, $no_pathselect_pages, $indices_all, $all_index_info, $indices_sorted, $completed_indices, 
+    global $esclient, $timezone, $esIndex, $no_pathselect_pages, $es_index_info, $all_index_info, $indices_sorted, $completed_indices, 
         $latest_completed_index, $fields, $indexinfo_updatetime, $index_starttimes, $indexinfo_expiretime, $index_spaceinfo;
 
     if (!isset($_SESSION['indices_uuids'])) {
@@ -260,21 +342,20 @@ function init()
     }
 
     // Get latest index info from ES
-    getIndexInfo();
+    $es_index_info = $esclient->getIndexInfo();
 
     // Set latest index info if force reload or index session info time expired
-    if (
-        isset($_GET['reloadindices']) || !isset($_SESSION['indexinfo']) ||
+    if (isset($_GET['reloadindices']) || !isset($_SESSION['indexinfo']) ||
         (isset($_SESSION['indexinfo']) && microtime(true) - $_SESSION['indexinfo']['update_time_ms'] > $indexinfo_expiretime)
     ) {
         list(
-            $indices_all, $all_index_info, $indices_sorted, $completed_indices, $latest_completed_index, 
+            $all_index_info, $indices_sorted, $completed_indices, $latest_completed_index, 
             $fields, $index_starttimes, $index_spaceinfo
         ) = setIndexInfo();
 
         if (isset($_SESSION['indexinfo'])) {
             // merge existing with new index info
-            $indices_all = array_merge($_SESSION['indexinfo']['indices_all'], $indices_all);
+            $es_index_info = array_merge($_SESSION['indexinfo']['es_index_info'], $es_index_info);
             $all_index_info = array_merge($_SESSION['indexinfo']['all_index_info'], $all_index_info);
             $indices_sorted = array_merge($_SESSION['indexinfo']['indices_sorted'], $indices_sorted);
             $completed_indices = array_merge($_SESSION['indexinfo']['completed_indices'], $completed_indices);
@@ -286,7 +367,7 @@ function init()
 
         $_SESSION['indexinfo'] = [
             'update_time_ms' => microtime(true),
-            'indices_all' => $indices_all,
+            'es_index_info' => $es_index_info,
             'all_index_info' => $all_index_info,
             'indices_sorted' => $indices_sorted,
             'completed_indices' => $completed_indices,
@@ -297,7 +378,7 @@ function init()
         ];
         $indexinfo_updatetime = $_SESSION['indexinfo']['update_time'] = new DateTime("now", new DateTimeZone($timezone));
     } else {
-        $indices_all = $_SESSION['indexinfo']['indices_all'];
+        $es_index_info = $_SESSION['indexinfo']['es_index_info'];
         $all_index_info = $_SESSION['indexinfo']['all_index_info'];
         $indices_sorted = $_SESSION['indexinfo']['indices_sorted'];
         $completed_indices = $_SESSION['indexinfo']['completed_indices'];
@@ -307,16 +388,6 @@ function init()
         $indexinfo_updatetime = $_SESSION['indexinfo']['update_time'];
         $index_spaceinfo = $_SESSION['indexinfo']['spaceinfo'];
     }
-
-    $indices_all = $_SESSION['indexinfo']['indices_all'];
-    $all_index_info = $_SESSION['indexinfo']['all_index_info'];
-    $indices_sorted = $_SESSION['indexinfo']['indices_sorted'];
-    $completed_indices = $_SESSION['indexinfo']['completed_indices'];
-    $latest_completed_index = $_SESSION['indexinfo']['latest_completed_index'];
-    $fields = $_SESSION['indexinfo']['fields'];
-    $index_starttimes = $_SESSION['indexinfo']['starttimes'];
-    $indexinfo_updatetime = $_SESSION['indexinfo']['update_time'];
-    $index_spaceinfo = $_SESSION['indexinfo']['spaceinfo'];
     
     // check for index in url
     if (isset($_GET['index']) && $_GET['index'] != "") {
@@ -346,14 +417,13 @@ function init()
 // remove index from globals and session vars
 function removeIndex($index, $uuid = null)
 {
-    global $indices_all, $all_index_info, $completed_indices;
+    global $all_index_info, $completed_indices;
 
     if (is_null($uuid)) {
         $uuid = array_search($index, $_SESSION['indices_uuids']);
     }
     unset($_SESSION['indices_uuids'][$uuid]);
     unset($_SESSION['indices_doccount'][$index]);
-    unset($_SESSION['indexinfo']['indices_all'][$index]);
     unset($_SESSION['indexinfo']['all_index_info'][$index]);
     if ($k = array_search($index, $_SESSION['indexinfo']['indices_sorted'])) {
         unset($_SESSION['indexinfo']['indices_sorted'][$k]);
@@ -363,9 +433,6 @@ function removeIndex($index, $uuid = null)
     }
     if ($k = array_search($index, $completed_indices)) {
         unset($completed_indices[$k]);
-    }
-    if ($k = array_search($index, $indices_all)) {
-        unset($indices_all[$k]);
     }
     if ($k = array_search($index, $all_index_info)) {
         unset($all_index_info[$k]);
@@ -377,7 +444,7 @@ function removeIndex($index, $uuid = null)
 // re-indexed will have a new uuid number
 function checkIndices()
 {
-    global $client, $esIndex, $completed_indices, $indices_info;
+    global $client, $esIndex, $completed_indices, $es_index_info;
 
     $indices_missing = $indices_changed = false;
 
@@ -394,7 +461,7 @@ function checkIndices()
                 $indices_missing = true;
             }
             // check uuid hasn't changed (from re-index)
-            if ($uuid !== $indices_info[$index]['uuid']) {
+            elseif ($uuid !== $es_index_info[$index]['uuid']) {
                 removeIndex($index, $uuid);
                 handleError('Index ' . $index . ' uuid changed', false, false, false);
                 $indices_changed = true;
@@ -415,6 +482,7 @@ function checkIndices()
     if ($indices_missing || $indices_changed) {
         deleteCookie('index');
         clearPaths();
+        createCookie('usecache', 0);
         if ($indices_missing) {
             $errmsg = "Selected indices are no longer available. Please select a different index.";
         } else {
@@ -456,21 +524,10 @@ function notifyNewIndex()
 }
 
 
-// get index info from ES and set globals
-function getIndexInfo()
-{
-    global $esclient, $indices_all, $indices_info;
-
-    $esclient->refreshIndices();
-    $indices_all = $esclient->getIndicesInfoCurl();
-    $indices_info = $esclient->getIndicesInfoCat();
-}
-
-
 // sets index global and session vars
 function setIndexInfo()
 {
-    global $client, $indices_all, $indices_info;
+    global $client, $es_index_info;
 
     $disabled_indices = array();
     $indices_sorted = array();
@@ -482,19 +539,18 @@ function setIndexInfo()
     $fields = array();
     $index_spaceinfo = array();
 
-    foreach ($indices_all as $key => $val) {
-        // get index uuid
-        $uuid = $val['settings']['index']['uuid'];
+    foreach ($es_index_info as $key => $val) {
+        // index uuid
+        $uuid = $val['uuid'];
 
         // skip index which we already have index uuid and doc count has not changed
-        if (
-            array_key_exists($uuid, $_SESSION['indices_uuids']) &&
-            $_SESSION['indices_doccount'][$key] == $indices_info[$key]['docs.count']
+        if (array_key_exists($uuid, $_SESSION['indices_uuids']) &&
+            $_SESSION['indices_doccount'][$key] == $val['docs_count']
         ) {
             continue;
         } else {
             $_SESSION['indices_uuids'][$uuid] = $key;
-            $_SESSION['indices_doccount'][$key] = $indices_info[$key]['docs.count'];
+            $_SESSION['indices_doccount'][$key] = $val['docs_count'];
             unset($_SESSION['toppath']);
         }
 
@@ -563,11 +619,11 @@ function setIndexInfo()
         }
 
         // add to indices_sorted using creation_date as key
-        $indices_sorted[$val['settings']['index']['creation_date']] = $key;
+        $indices_sorted[$val['creation_date']] = $key;
 
         // Get size of index
         // convert to bytes
-        $indexsize = $indices_info[$key]['store.size'];
+        $indexsize = $val['store_size'];
         if (strpos($indexsize, 'gb')) {
             $indexsize = str_replace('gb', '', $indexsize);
             $indexsize = $indexsize * 1024 * 1024 * 1024;
@@ -595,7 +651,9 @@ function setIndexInfo()
         try {
             $queryResponse = $client->search($searchParams);
         } catch (Missing404Exception $e) {
-            handleError("Selected indices are no longer available.");
+                deleteCookie('index');
+                clearPaths();
+                handleError("Selected indices are no longer available. Please select a different index.");
         } catch (Exception $e) {
             handleError('ES error: ' . $e->getMessage(), true);
         }
@@ -603,45 +661,26 @@ function setIndexInfo()
         foreach ($queryResponse['hits']['hits'] as $hit) {
             $index_spaceinfo[$key][$hit['_source']['path']] = $hit['_source'];
         }
+
+        // add all fields from index mappings to fields_all
+        foreach ($val['fields'] as $field) {
+            if (!in_array($field, $fields)) {
+                $fields[] = $field;
+            }
+        }
     }
 
     // sort completed indices by creation_date
     krsort($indices_sorted);
 
-    // fields to not display
-    $field_exclusions = [
-        'path',
-        'total',
-        'used',
-        'free',
-        'available',
-        'file_size',
-        'file_size_du',
-        'file_count',
-        'dir_count',
-        'start_at',
-        'end_at',
-        'crawl_time',
-        'diskover_ver'
-    ];
-
     // get completed indices
-    // get all latest indices based on index's top paths
-    // get all index fields for file/directory docs
+    // get all latest indices based on index's top path
     foreach ($indices_sorted as $key => $val) {
         if (!in_array($val, $disabled_indices)) {
             if (is_null($latest_completed_index)) {
                 $latest_completed_index = $val;
             }
             $completed_indices[] = $val;
-            // add all fields from index mappings to fields_all
-            $index_fields = $indices_all[$val]['mappings']['properties'];
-            foreach ($index_fields as $k => $v) {
-                // exclude certain fields which are not in file/directory docs
-                if (!in_array($k, $field_exclusions) && !in_array($k, $fields)) {
-                    $fields[] = $k;
-                }
-            }
         }
     }
 
@@ -653,7 +692,7 @@ function setIndexInfo()
     asort($fields);
 
     return [
-        $indices_all, $all_index_info, $indices_sorted, $completed_indices, $latest_completed_index, 
+        $all_index_info, $indices_sorted, $completed_indices, $latest_completed_index, 
         $fields, $index_starttimes, $index_spaceinfo
     ];
 }
@@ -664,7 +703,7 @@ function get_es_path($index, $numofpaths)
     /* try to get a top level path from ES index
     if there are multiple paths, return them all, 
     don't return if the path is still being indexed */
-    global $client, $client;
+    global $client;
 
     $searchParams['body'] = [];
 
@@ -681,6 +720,10 @@ function get_es_path($index, $numofpaths)
     // Send search query to Elasticsearch and get results
     try {
         $queryResponse = $client->search($searchParams);
+    } catch (Missing404Exception $e) {
+        deleteCookie('index');
+        clearPaths();
+        handleError("Selected indices are no longer available. Please select a different index.");
     } catch (Exception $e) {
         handleError('ES error: ' . $e->getMessage(), true);
     }
@@ -845,7 +888,7 @@ function deleteCookie($cname)
 // saved search query functions
 function saveSearchQuery($req)
 {
-    if (!$_REQUEST['userinput']) return;
+    if (!isset($_REQUEST['userinput']) || !$_REQUEST['userinput']) return;
     
     ($req === "") ? $req = "type:(file OR directory)" : "";
     if (getCookie('savedsearches') != "") {
@@ -1027,7 +1070,7 @@ function writeToFile($fp, $file, $data) {
 // sort search results and set cookies for search pages
 function sortSearchResults($request, $searchParams)
 {
-    if (!$request['sort'] && !$request['sort2'] && !getCookie("sort") && !getCookie("sort2")) {
+    if (!isset($request['sort']) && !isset($request['sort2']) && !getCookie("sort") && !getCookie("sort2")) {
         if (getCookie('unsorted') == 1) {
             $searchParams['body']['sort'] = [];
         } else {
@@ -1042,10 +1085,10 @@ function sortSearchResults($request, $searchParams)
         $sortarr = ['sort', 'sort2'];
         $sortorderarr = ['sortorder', 'sortorder2'];
         foreach ($sortarr as $key => $value) {
-            if ($request[$value] && !$request[$sortorderarr[$key]]) {
+            if (isset($request[$value]) && !$request[$sortorderarr[$key]]) {
                 $searchParams['body']['sort'] = $request[$value];
                 createCookie($value, $request[$value]);
-            } elseif ($request[$value] && $request[$sortorderarr[$key]]) {
+            } elseif (isset($request[$value]) && $request[$sortorderarr[$key]]) {
                 array_push($searchParams['body']['sort'], [$request[$value] => ['order' => $request[$sortorderarr[$key]]]]);
                 createCookie($value, $request[$value]);
                 createCookie($sortorderarr[$key], $request[$sortorderarr[$key]]);